pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiovantage"
dynamic = ["version"]
description = "Interact with and control Vantage InFusion home automation controllers."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "James Smith", email = "james@loopj.com"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "attrs==23.1.0",
    "xsdata==23.8",
    "xsdata_attrs==23.8",
    "typing_extensions>=4.6.3,<5.0",
]

[project.urls]
Documentation = "https://github.com/loopj/aiovantage#readme"
Issues = "https://github.com/loopj/aiovantage/issues"
Source = "https://github.com/loopj/aiovantage"

[project.optional-dependencies]
dev = [
    "black==23.7.0",
    "mypy==1.5.0",
    "ruff==0.0.284",
    "pre-commit==3.3.3",
]

[tool.hatch.version]
path = "src/aiovantage/__about__.py"

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11"]

[tool.hatch.envs.dev]
features = ["dev"]

[tool.hatch.envs.lint]
features = ["dev"]

[tool.hatch.envs.lint.scripts]
style = [
  "ruff {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff --fix {args:.}",
]
typing = "mypy {args:.}"

[tool.mypy]
strict = true

[tool.black]
target-version = ["py38"]
skip-string-normalization = true

[tool.ruff]
target-version = "py38"
select = [
    "F",    # pyflakes
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "D",    # pydocstyle
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "Q",    # flake8-quotes
    "C90",  # mccabe
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.per-file-ignores]
"src/aiovantage/config_client/models/**/*.py" = [
    "D106"  # Missing docstring in public nested class
]

[tool.ruff.isort]
known-first-party = ["aiovantage"]

[tool.ruff.mccabe]
max-complexity = 15

[tool.ruff.pydocstyle]
convention = "google"